refactor: extract the scrubbing rules into lib/scrub.ts
telemetry.ts statically imports @sentry/react, and ES imports are hoisted, so any
module importing a redaction helper from it pulls the SDK in regardless of DSN. Move
the pure rules (REDACTED, isSensitiveKey, stripQueryString, scrubUrlsIn, scrubDeep)
into a new lib/scrub.ts that imports nothing, so a later feedback trail can reuse them
without depending on error tracking. telemetry.ts re-exports the four public names and
keeps a local `scrub` alias so the rest of the file is unchanged.
