+++ apps/web/src/lib/telemetry.ts
@@ -1,63 +1,11 @@
-// Redaction rules for anything sent to the error tracker.
-//
-// SYNC OBLIGATION: this deny-list exists in three places and they must stay in step:
-//   - src/Diariz.Worker/telemetry.py
-//   - src/Diariz.Api/Services/SentryScrubber.cs
-//   - this file
-// Each has a test asserting the shared set below. Those tests catch a REMOVAL from one copy, but
-// they cannot catch an ADDITION made to only one copy - if you add an entry here, add it to the
-// other two and to their tests.
+import { REDACTED, isSensitiveKey, stripQueryString, scrubUrlsIn, scrubDeep } from "./scrub";
 
-export const REDACTED = "[redacted]";
+// Local alias: the rest of this file calls the walker `scrub`.
+const scrub = scrubDeep;
 
-// Exact field names that carry meeting content.
-const DENY_EXACT = new Set([
-  "text", "transcript", "transcription", "segments", "words", "summary",
-  "minutes", "note", "notes", "content", "authorization", "cookie", "cookies",
-  // ECAPA voiceprint vectors (biometric data identifying a speaker by voice).
-  "embedding", "embeddings",
-]);
-
-// Substrings marking a credential regardless of the surrounding name.
-const DENY_SUBSTRING = ["secret", "token", "password", "api_key", "apikey", "access_key", "accesskey"];
-
-/** True when a field with this name must never leave the browser. */
-export function isSensitiveKey(key: string): boolean {
-  const lowered = String(key).toLowerCase();
-  if (DENY_EXACT.has(lowered)) return true;
-  return DENY_SUBSTRING.some((marker) => lowered.includes(marker));
-}
-
-/**
- * Drop a URL's query string, keeping its path.
- *
- * Load-bearing, not cosmetic: @microsoft/signalr appends `?access_token=<JWT>` to the hub's negotiate
- * and WebSocket URLs, because a browser cannot set an Authorization header on a WS handshake. A
- * key-name deny-list cannot reach that - a query string is one opaque value, not a named field. The
- * path is kept because it is diagnostically useful and carries no credential.
- *
- * Anything that is not a URL is returned unchanged. `null`/`undefined` coerce to `""` (matching
- * isSensitiveKey's String() coercion above) rather than throwing - Task 3 wires this to Sentry
- * fields typed `string | undefined`, and a throw here would happen inside a beforeSend hook, taking
- * the whole scrub pipeline down at exactly the moment something has already gone wrong.
- */
-export function stripQueryString(value: string): string {
-  const raw = String(value ?? "");
-  const cut = raw.indexOf("?");
-  if (cut === -1) return raw;
-  // Only treat it as a URL if what precedes the "?" looks like one; otherwise leave free text alone.
-  const head = raw.slice(0, cut);
-  const isUrl = /^https?:\/\//i.test(head) || head.startsWith("/");
-  return isUrl ? head : raw;
-}
-
-/**
- * Apply stripQueryString to every whitespace-separated token, for descriptions like "GET <url>".
- * `null`/`undefined` coerce to `""` for the same reason as stripQueryString above.
- */
-export function scrubUrlsIn(text: string): string {
-  return String(text ?? "").split(" ").map(stripQueryString).join(" ");
-}
+// Re-exported so any existing importer of this file still resolves; the definitions themselves now
+// live in ./scrub - see the SYNC OBLIGATION header there.
+export { REDACTED, isSensitiveKey, stripQueryString, scrubUrlsIn };
 
 // Named imports (rather than `import * as Sentry`) so a bundler can tree-shake the integrations this
 // app never enables (session replay, user feedback, tracing) - referencing the whole namespace object
@@ -159,34 +107,6 @@ function scrubAttributes<T>(data: T, seen: WeakSet<object> = new WeakSet(), dept
     return out as T;
   } finally {
     seen.delete(data);
-  }
-}
-
-/**
- * Recursively redact sensitive values by field name. Pure: returns a new structure, never mutates the
- * input.
- *
- * Cycle-guarded for the same reason `scrubAttributes` is, and it is not theoretical: before the guard
- * existed, handing `beforeBreadcrumb` a self-referencing `data` object threw
- * `RangeError: Maximum call stack size exceeded` out of THIS function - breadcrumb data is arbitrary
- * app-supplied structure, and a throw from inside a breadcrumb hook is worse than the leak the
- * recursion was added to close, because it takes the page with it.
- */
-function scrub<T>(value: T, seen: WeakSet<object> = new WeakSet(), depth = 0): T {
-  if (!isWalkable(value)) return value;
-  if (seen.has(value) || depth >= MAX_DEPTH) return CIRCULAR as unknown as T;
-  seen.add(value);
-  try {
-    if (Array.isArray(value)) {
-      return value.map((item) => scrub(item, seen, depth + 1)) as unknown as T;
-    }
-    const out: Record<string, unknown> = {};
-    for (const [k, v] of Object.entries(value as Record<string, unknown>)) {
-      out[k] = isSensitiveKey(k) ? REDACTED : scrub(v, seen, depth + 1);
-    }
-    return out as unknown as T;
-  } finally {
-    seen.delete(value);
   }
 }
