+++ src/Diariz.Worker/telemetry.py
@@ -25,7 +25,7 @@ REDACTED = "[redacted]"
 # ---- Cross-runtime deny-list ------------------------------------------------------------------
 # KEEP IN SYNC WITH:
 #   - src/Diariz.Api/Services/SentryScrubber.cs (DenyExact / DenySubstring)
-#   - apps/web/src/lib/telemetry.ts (DENY_EXACT / DENY_SUBSTRING)
+#   - apps/web/src/lib/scrub.ts (DENY_EXACT / DENY_SUBSTRING)
 # The worker, the API and the browser SPA report to the same GlitchTip instance and must redact the
 # same names. These lists HAVE silently diverged before: "embedding"/"embeddings" (the ECAPA
 # voiceprint vectors) were added here only - even though the API is the runtime that actually stores
@@ -34,7 +34,7 @@ REDACTED = "[redacted]"
 # What the guard-rail is, honestly: each runtime has a test that pins this shared set
 # (tests/test_telemetry.py::test_the_shared_cross_runtime_deny_list_is_covered here,
 # SentryScrubberTests.IsSensitiveKey_CoversEveryNameInTheCrossRuntimeDenyList there, and
-# telemetry.test.ts's isSensitiveKey suite in the browser). That catches a REMOVAL from any one
+# scrub.test.ts's isSensitiveKey suite in the browser). That catches a REMOVAL from any one
 # copy. It cannot catch an ADDITION made to only one copy - three tests in three languages have no
 # shared source of truth - so adding a name here means adding it to the other two, and to their
 # tests.
